[ND0xxx] Changed pow to powf for type consistency

+++ src/shared/sensors/ND015X/ND015A.cpp
@@ -187,8 +187,8 @@ ND015XData ND015A::sampleImpl()
     }
 
     float normalizedPressure =
-        (static_cast<float>(spiDataIn) - 0.05f * pow(2, 16)) /
-        (0.9f * pow(2, 16));
+        (static_cast<float>(spiDataIn) - 0.05f * powf(2, 16)) /
+        (0.9f * powf(2, 16));
 
     data.pressure =
         normalizedPressure * ND015A::FULLSCALE * Constants::PSI_TO_PASCAL +
